Extract Navigator Screen Options From NFT search bar (#2407)
