+++ src/Search/SearchHeader.tsx
@@ -1,9 +1,48 @@
+import {StackNavigationOptions} from '@react-navigation/stack'
 import React from 'react'
+import {StyleSheet, TouchableOpacity, TouchableOpacityProps} from 'react-native'
 
+import {Icon} from '../components'
 import {SearchBar} from './SearchBar'
 import {useSearch} from './SearchContext'
 
-export interface Props {
+export const useSearchHeaderOptions = ({placeHolderText, title}) => {
+  const [searchVisible, setSearchVisible] = React.useState(false)
+  const {clearSearch} = useSearch()
+  const handleSearchClose = () => {
+    setSearchVisible(false)
+    clearSearch()
+  }
+
+  const searchHeaderOptions: StackNavigationOptions = searchVisible
+    ? {
+        headerTitleContainerStyle: styles.headerTitleContainer,
+        headerTitle: () => <SearchHeader placeholder={placeHolderText} onClose={handleSearchClose} />,
+        headerLeftContainerStyle: styles.disableFlex,
+        headerLeft: () => null,
+        headerRightContainerStyle: styles.disableFlex,
+        headerRight: () => null,
+      }
+    : {
+        headerTitleContainerStyle: styles.headerTitleContainer,
+        title: title,
+        headerRight: () => <SearchButton onPress={() => setSearchVisible(true)} />,
+        headerRightContainerStyle: {
+          ...styles.disableFlex,
+          paddingHorizontal: 16,
+          position: 'absolute',
+          right: 0,
+          bottom: 0,
+          top: 0,
+        },
+        headerLeftContainerStyle: styles.disableFlex,
+        headerLeft: () => null,
+      }
+
+  return {searchHeaderOptions}
+}
+
+interface Props {
   placeholder: string
   onClose?(): void
 }
@@ -21,3 +60,24 @@ export const SearchHeader = ({placeholder, onClose}: Props) => {
     />
   )
 }
+
+const SearchButton = (props: TouchableOpacityProps) => (
+  <TouchableOpacity {...props} hitSlop={{top: 100, left: 100, right: 100, bottom: 100}}>
+    <Icon.Magnify size={26} />
+  </TouchableOpacity>
+)
+
+const styles = StyleSheet.create({
+  disableFlex: {
+    flex: undefined,
+    flexGrow: undefined,
+    flexBasis: undefined,
+    flexShrink: undefined,
+  },
+  headerTitleContainer: {
+    flex: 1,
+    marginHorizontal: undefined,
+    maxWidth: undefined,
+    alignItems: 'center',
+  },
+})
